Should fix the backfill bug
Aimed at fixing the bug in which a queue repop results in the match never ending.

Calling allPlayers in this line should let any player in the match interact with GG instead of just the ones who were backfilled in the repop. This was probably the main cause of that bug.