I did everything that bigboot asked
I also was trying to see if there was a way to put lines 217-240 and 268-288 is one function but kotlin was saying that I cant just return a addEmbedCompat. I also was thinking about making the 2 parent handleInteraction into one function but did not really know how to do so without the two classes being the same.